Check for enum class type when building attributes
A bit of an edge case perhaps, but this fixes some strange side effects seen when using enums that have an __invoke() method, causing is_callable() to return true.
This happens to be the case for a few popular enum packages out there.

We correct this by checking if the factory attribute is an enum, and make sure we do not treat it as a callable.